Add support for vscode-textmate in the browser
