+++ package.json
@@ -41,6 +41,7 @@
     "native-keymap": "1.2.5",
     "native-watchdog": "1.0.0",
     "node-pty": "0.9.0-beta9",
+    "onigasm-umd": "^2.2.2",
     "semver": "^5.5.0",
     "spdlog": "0.8.1",
     "sudo-prompt": "8.2.0",
@@ -51,7 +52,7 @@
     "vscode-proxy-agent": "0.4.0",
     "vscode-ripgrep": "^1.2.5",
     "vscode-sqlite3": "4.0.7",
-    "vscode-textmate": "^4.0.1",
+    "vscode-textmate": "^4.1.1",
     "vscode-xterm": "3.14.0-beta4",
     "yauzl": "^2.9.1",
     "yazl": "^2.4.3"
