+++ src/vs/code/browser/workbench/workbench.html
@@ -14,6 +14,7 @@
 	<script>
 		self.CONNECTION_AUTH_TOKEN = '{{CONNECTION_AUTH_TOKEN}}';
 		self.USER_HOME_DIR = '{{USER_HOME_DIR}}';
+		self.SERVER_APP_ROOT = '{{SERVER_APP_ROOT}}';
 	</script>
 
 	<!-- Startup via workbench.js -->
