+++ src/vs/workbench/services/textMate/common/cgmanifest.json
